fix createPost and architect
