+++ src/main/java/com/hanu/mediumclone/controllers/PostController.java
@@ -4,6 +4,7 @@ import java.util.List;
 
 import javax.validation.Valid;
 
+import com.hanu.mediumclone.services.PostService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -14,7 +15,6 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.hanu.mediumclone.models.Post;
-import com.hanu.mediumclone.services.PostService;
 
 @RestController
 @RequestMapping(value = "/post")
@@ -31,7 +31,7 @@ public class PostController {
 	@PostMapping("/create")
     public ResponseEntity<Post> createPost(@RequestBody @Valid Post post) {
 		Post createdPost = postService.createPost(post);
-		return new ResponseEntity<Post>(createdPost, HttpStatus.OK);
+		return new ResponseEntity<>(createdPost, HttpStatus.OK);
     }
 
 }
