+++ src/main/resources/application.properties
@@ -1,5 +1,5 @@
 ## Database Properties
-spring.datasource.url = jdbc:mysql://localhost:3306/medium-clone?useSSL=false
+spring.datasource.url = jdbc:mysql://localhost:3306/medium-clone?useSSL=false&allowPublicKeyRetrieval=true
 spring.datasource.username = root
 spring.datasource.password = root
